backend/requirements.txt
# ── Core ──────────────────────────────────────
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ── Database ──────────────────────────────────
sqlalchemy>=2.0.23
alembic>=1.13.0
aiosqlite>=0.19.0
# asyncpg>=0.29.0  # uncomment for PostgreSQL in production

# ── HTTP / LLM ───────────────────────────────
httpx>=0.25.1

# ── CSV / File handling ──────────────────────
chardet>=5.2.0
python-multipart>=0.0.6

# ── Auth / Security ──────────────────────────
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0

# ── Development / Testing ────────────────────
pytest>=7.4.3
pytest-asyncio>=0.21.1
coverage>=7.3.0

